Add a category and implement recommendation

+++ server/controller/shoppingList/DeleteShoppingItem.php
@@ -1,7 +1,10 @@
 <?php
 require_once("../../model/ShoppingListItem.php");
+require_once("../../DAO/PostShoppingListDAO.php");
 require_once ("../../DAO/shoppingListDAO.php");
+session_start();
 
+$poststate = false;
 $status = false;
 $result = [];
 if(isset($_GET['id'])){
@@ -10,7 +13,20 @@ if(isset($_GET['id'])){
     var_dump($id);
 
     $shoppingListDAO = new shoppingListDAO();
-    $status = $shoppingListDAO->deleteShoppingListItem($id);
+    $PostShoppingListDAO = new PostShoppingListDAO;
+    $getItems = $shoppingListDAO->getShoppingItembyitemID($id);
+    foreach($getItems as $getItem){
+        $item = $getItem->getItem();
+        $category = $getItem->getCategory();
+        $quantity = $getItem->getQuantity();
+        $userID = $getItem->getUserID();
+        $poststate = $PostShoppingListDAO->insertPostShoppingListItem($item, $category, $quantity, $userID);
+
+    }
+    if ($poststate){
+        $status = $shoppingListDAO->deleteShoppingListItem($id);
+    }
+   
 }
 
 if ($status)
